+++ outputs.tf
@@ -15,12 +15,12 @@
  */
 
 output "network_name" {
-  value       = google_compute_network.network.name
+  value       = data.google_compute_network.network.name
   description = "The name of the VPC being created"
 }
 
 output "network_self_link" {
-  value       = google_compute_network.network.self_link
+  value       = data.google_compute_network.network.self_link
   description = "The URI of the VPC being created"
 }